Update active-directory-azureadjoin-devices-group-policy.md
We ran into a problem when using useraccounts with memberships in a group that had a rID of 7515 were unable to log on to Microsoft Online resources. Changing the three Issuance Transform Rules for Value =~ "515$" to Value =~ "-515$" fixes the problem, because then it actually looks for the sID to end with –515, eliminating all groups ending with 515, like 1515, 2515, 3515, 4515, 5515, 6515, 7515, etc.